pyproject.toml
[project]
name = "slither-detector-101"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "slither-analyzer>=0.11.5",
]
